+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 0.0.3
+- Use default injector for injected & registered if none specified
+
 ## 0.0.2
 - Renamed library Injector -> Discoverer
 - Configured CI & badges